Remove pointless getNextJobFromQueue abstraction
getNextJobFromQueue was thought of abstracting the extraction of jobs
from the jobQueue. Although it was useful for encapsulating the decision
 of which job is next, it added an indirection, increasing the
complexity of the logic. As for the the job scheduling is fairly simple,
and until there are no other scheduling policies in place, this
abstraction added more drawbacks than benefits.